Stop using ament_target_dependencies. (#86)
We are slowly moving away from its use, so stop using it
here.  While we are in here, notice some things that makes
this easier:

1. pluginlib is absolutely a public dependency of this package.
Because of that, we can just rely on the PUBLIC export of it,
and we don't need to link it into every test.  But that also means
we don't need some of the forward-declarations that were in
loader_fwds.hpp, as we can just get those through the header file.
2. republish.hpp doesn't really need to exist at all.  That's
because it is only a header file, but the implementation is in
an executable.  Thus, no downstream could ever use it.  So just
remove the file and put the declaration straight into the cpp file.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>